[CSL-1606] Add termination policy to `onNewSlot`
`onNewSlot` worker now takes `ActionTerminationPolicy` as an argument.
Parameters are wrapped in `OnNewSlotParams` type.
`ActionTerminationPolicy` allows to cancel running action when new slot starts.
This feature is used in two workers: block creator and `Update.processNewSlot`.
The detailed analysis was performed here:
https://github.com/input-output-hk/cardano-sl/pull/1163#issuecomment-318136810
In particular it allowed us to avoid `fork` in block creator.
Ideally we also want to separate actual block creation and network stuff which
happens afterwards, but I think it's minor, so it hasn't been done.